migrate more tests to datachain
