+++ tests/func/test_dataset_query.py
@@ -948,81 +948,6 @@ def test_class_udf(cloud_test_catalog, batch):
     [("s3", True)],
     indirect=True,
 )
-@pytest.mark.xdist_group(name="tmpfile")
-def test_udf_reuse_on_error(cloud_test_catalog_tmpfile):
-    catalog = cloud_test_catalog_tmpfile.catalog
-    sources = [cloud_test_catalog_tmpfile.src_uri]
-    globs = [s.rstrip("/") + "/*" for s in sources]
-    catalog.index(sources)
-    catalog.create_dataset_from_sources("animals", globs, recursive=True)
-
-    error_state = {"error": True}
-
-    @udf((C.path,), {"path_len": Int})
-    def name_len_maybe_error(path):
-        if error_state["error"]:
-            # A udf that raises an exception
-            raise RuntimeError("Test Error!")
-        return (len(path),)
-
-    q = (
-        DatasetQuery(name="animals", version=1, catalog=catalog)
-        .filter(C.size < 13)
-        .filter(C.path.glob("cats*") | (C.size < 4))
-        .add_signals(name_len_maybe_error)
-        .select(C.path, C.path_len)
-    )
-    with pytest.raises(RuntimeError, match="Test Error!"):
-        q.db_results()
-
-    # Simulate fixing the error
-    error_state["error"] = False
-
-    # Retry Query
-    result = q.db_results()
-
-    assert len(result) == 3
-    for r in result:
-        # Check that the UDF ran successfully
-        assert len(r[0]) == r[1]
-
-
-@pytest.mark.parametrize(
-    "cloud_type,version_aware",
-    [("s3", True)],
-    indirect=True,
-)
-@pytest.mark.xdist_group(name="tmpfile")
-def test_udf_parallel_interrupt(cloud_test_catalog_tmpfile, capfd):
-    catalog = cloud_test_catalog_tmpfile.catalog
-    sources = [cloud_test_catalog_tmpfile.src_uri]
-    globs = [s.rstrip("/") + "/*" for s in sources]
-    catalog.index(sources)
-    catalog.create_dataset_from_sources("animals", globs, recursive=True)
-
-    @udf(("path",), {"name_len": Int})
-    def name_len_interrupt(_name):
-        # A UDF that emulates cancellation due to a KeyboardInterrupt.
-        raise KeyboardInterrupt
-
-    q = (
-        DatasetQuery(name="animals", version=1, catalog=catalog)
-        .filter(C.size < 13)
-        .filter(C.path.glob("cats*") | (C.size < 4))
-        .add_signals(name_len_interrupt, parallel=-1)
-    )
-    with pytest.raises(RuntimeError, match="UDF Execution Failed!"):
-        q.db_results()
-    captured = capfd.readouterr()
-    assert "KeyboardInterrupt" in captured.err
-    assert "semaphore" not in captured.err
-
-
-@pytest.mark.parametrize(
-    "cloud_type,version_aware",
-    [("s3", True)],
-    indirect=True,
-)
 @pytest.mark.parametrize("batch", [False, True])
 @pytest.mark.parametrize("workers", (1, 2))
 @pytest.mark.skipif(
